requirements.txt
torch
torchvision
numpy
matplotlib
scikit-learn
umap-learn
pillow